# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant Booker API",
    "version": "1.0.0",
    "description": "API for managing restaurants"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ]
            }
          }
        },
        "required": [
          "error",
          "details"
        ]
      },
      "Restaurant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "cuisineType": {
            "type": "string",
            "enum": [
              "ASIAN",
              "PIZZA",
              "AMERICAN",
              "MEXICAN"
            ]
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "rating": {
            "type": "number"
          },
          "capacity": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "capacity"
        ]
      },
...
      "Favourite": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "restaurantId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FavouriteWithRestaurant": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Favourite"
          },
          {
            "type": "object",
            "properties": {
              "restaurant": {
                "$ref": "#/components/schemas/Restaurant"
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/me/favourites": {
      "get": {
        "summary": "Get the authenticated user's favourite restaurants",
        "tags": [
          "Favourites"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of favourite restaurants.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FavouriteWithRestaurant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/favourites/{restaurantId}": {
      "post": {
        "summary": "Add a restaurant to favourites",
        "tags": [
          "Favourites"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "restaurantId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Restaurant added to favourites.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favourite"
                }
              }
            }
          },
          "400": {
            "description": "Invalid restaurant id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Restaurant not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove a restaurant from favourites",
        "tags": [
          "Favourites"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "restaurantId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Restaurant removed from favourites."
          },
          "400": {
            "description": "Invalid restaurant id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Favourite not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}